requirements.txt
django
djangorestframework
psycopg2
gunicorn
numpy